ZenKit asset library, big refactor
General:
- moved windows resources (window icon, manifest, etc.) into subfolder viewer/win_resources
- reduced BufferSize memmbers to uint16_t so BufferSize can be used for texture info etc. without issues
- add launch.vs.json to Git to make launching from IDE easier

Shaders:
- remove DirectX SDK (June 2010) and any uses of it to make it possible to use Windows SDK shader compilation methods
  -> switch to "D3DCompileFromFile" from "D3DX11CompileFromFileW"
- compilation flags: enable shader optimization in release mode and strictness in all modes
- implement compute shader creation

Windows headers:
- D3D release function in Dx.h is now templated on any class so it can be called without including WinDx.h headers
  -> enforcing that passed types are actually subclass of IUnknown is done with static_assert in impl
- replaced HWND window handle type with void* to allow Renderer.h to no longer pull in windows headers
- general reduction of windows header inclusion in several files

Primitives/Math:
- CPP concepts have been introduced to describe common base structs like vectors or color (current concepts: RGBA, Vec2, Vec3)
- most util functions are now generic over actual vector/color structs thanks to those concepts

D3D Buffers:
- new namespace render::d3d specifically for D3D helper functions
  -> moved/created functions for buffer creation and updating (texture2D, constant, vertex, structured buffers)
  -> moved/created functions for SRV/UAV creation and readback buffers

ZenKit:
- add ZenKit as alternative asset laoding library
- remove libsquish build from ZenLib due to build conflicts with ZenKit
- move ZenLib/ZenKit independent code from ZenLoader to new file VobLoader
  - extract main VOB lighting code into its own function in VobLoader from ZenLoader # loadVobs
- for ZenLib specific funtions create equivalent ZenKit functions or abstract over both
  - MeshFromVdfLoader and AssetCache now have two variants of all functions
  - AssetFinder.cpp has internal flag that switches library usage between ZenKit and ZenLib
  - directly exposing VFS is now deprecated, AssetFinder is now mostly transparent regarding which lib is used
  - ZenKitLoader is ZenKit equivalent to ZenLoader

ZenLib:
This commit still contains working code for using only ZenLib to load all assets. ZenLib will be removed later.
However, usage of ZenLib has been changed a lot to maximize code reuse and comparability between ZenLib and ZenKit.
- packed meshes are no longer used, raw asset data is loaded directly instead
  -> MeshFromVdfLoader has been effectily entirely re-implemented
  -> AssetCache has become simpler since packed meshes no longer need to be cached

Asset Loading:
- new struct FileHandle abstracts over any asset source (file path, VFS)
- new struct FileData abstracts over any raw asset file content (from which stream of bytes can be retrieved) regardless of ownership
  -> mmaped file (owned), buffer (owned), buffer view (not owned)
- MeshFromVdfLoader rewrite
  - share more code between level/vobmesh loading
  - extract lightmap uv calculations
- ZenKit loader now correctly loads MDH/MDM asset pairs, which was never implemented correctly for ZenLib

+++ src/viewer/GameLoop.h
@@ -4,8 +4,8 @@
 
 namespace viewer
 {
-	void init(HWND hWnd, Arguments args, uint32_t width, uint32_t height);
-	void onWindowResized(uint32_t width, uint32_t height);
+	void init(HWND hWnd, Arguments args, uint16_t width, uint16_t height);
+	void onWindowResized(uint16_t width, uint16_t height);
 	void onWindowDpiChanged(float dpiScale);
 	void execute();
 	void cleanup();
